Refactor - Rename according to new design - Create useDiscover hook, added join token related code - Move agent related components and variables to its specific file - Remove unecessary checking for null context
