+++ packages/teleport/src/Discover/Discover.tsx
@@ -17,80 +17,35 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
 import styled from 'styled-components';
-import {
-  Flex,
-  ButtonPrimary,
-  ButtonSecondary,
-  Text,
-  Box,
-  Indicator,
-} from 'design';
+import { Flex, ButtonPrimary, Text, Box, Indicator } from 'design';
 import TopNavUserMenu from 'design/TopNav/TopNavUserMenu';
 import { Danger } from 'design/Alert';
 import * as Icons from 'design/Icon';
-import useAttempt from 'shared/hooks/useAttemptNext';
 import { MenuItem, MenuItemIcon } from 'shared/components/MenuAction';
 import cfg from 'teleport/config';
 import { useDiscoverContext } from './discoverContextProvider';
+import { useDiscover, State, AgentStep } from './useDiscover';
+import { agentStepTitles, agentViews, SelectAgent } from './AgentConnect';
 
-// agentStepTitles defines the titles per steps defined by enum `AgentStep`.
-//
-// We use the enum `AgentStep` numerical values to access the list's value,
-// so this list's order and length must be equal to the enum.
-const agentStepTitles: string[] = [
-  'Requirement Gathering',
-  'Teleport Installation',
-  'Agent Initiation',
-  'Role Configuration',
-  'Connect Resource',
-  'Test Resource',
-];
+export default function Container() {
+  const ctx = useDiscoverContext();
+  const state = useDiscover(ctx);
 
-// agentViews defines the list of components required per AgentKind per steps
-// defined by enum `AgentStep`.
-//
-// We use the enum `AgentStep` numerical values to access the list's value,
-// so the list's order and length must be equal to the enum.
-const agentViews: Record<AgentKind, AgentStepComponent[]> = {
-  node: [RequirementGatheringNode, TeleportInstallation, AgentInitiation],
-  app: [RequirementGathering, TeleportInstallation, AgentInitiation],
-};
+  return <Discover {...state} />;
+}
 
-export function Discover() {
-  const discoverCtx = useDiscoverContext();
-
-  const [user, setUser] = React.useState('');
-  const [currentStep, setCurrentStep] = React.useState<AgentStep>(0);
-  const [agentKind, setAgentKind] = React.useState<AgentKind>();
-  const [agentMeta, setAgentMeta] = React.useState<AgentMeta>();
-
-  const { attempt: initAttempt, run: initRun } = useAttempt('processing');
-
-  function onSelectResource(kind: AgentKind) {
-    setAgentKind(kind);
-  }
-
-  function nextStep() {
-    setCurrentStep(currentStep + 1);
-  }
-
-  function prevStep() {
-    setCurrentStep(currentStep - 1);
-  }
-
-  function updateResourceMeta(meta: AgentMeta) {
-    setAgentMeta(meta);
-  }
-
-  React.useEffect(() => {
-    initRun(() =>
-      discoverCtx.userService.fetchUserContext().then(c => setUser(c.username))
-    );
-  }, []);
-
+export function Discover({
+  initAttempt,
+  user,
+  currentStep,
+  selectedAgentKind,
+  logout,
+  onSelectResource,
+  ...agentProps
+}: State) {
   let AgentComponent;
-  if (agentKind) {
-    AgentComponent = agentViews[agentKind][currentStep];
+  if (selectedAgentKind) {
+    AgentComponent = agentViews[selectedAgentKind][currentStep];
   }
 
   return (
@@ -105,20 +60,17 @@ export function Discover() {
       )}
       {initAttempt.status === 'success' && (
         <>
-          <TopBar onLogout={discoverCtx.logout} user={user} />
+          <TopBar onLogout={logout} user={user} />
           <Flex p={5} alignItems="flex-start">
             <SideNavAgentConnect currentStep={currentStep} />
             <Box width="100%" height="100%" minWidth="0">
-              {currentStep === AgentStep.RequirementGather && (
-                <SelectAgent onClick={onSelectResource} agentKind={agentKind} />
-              )}
-              {AgentComponent && (
-                <AgentComponent
-                  nextStep={nextStep}
-                  prevStep={prevStep}
-                  updateResourceMeta={updateResourceMeta}
+              {currentStep === AgentStep.Select && (
+                <SelectAgent
+                  onClick={onSelectResource}
+                  agentKind={selectedAgentKind}
                 />
               )}
+              {AgentComponent && <AgentComponent {...agentProps} />}
             </Box>
           </Flex>
         </>
@@ -153,7 +105,7 @@ function TopBar(props: { onLogout: VoidFunction; user: string }) {
       border="1px solid"
     >
       <Text typography="h4" bold>
-        Getting Started
+        Access Management
       </Text>
       <TopNavUserMenu
         menuListCss={() => `
@@ -175,46 +127,6 @@ function TopBar(props: { onLogout: VoidFunction; user: string }) {
         </MenuItem>
       </TopNavUserMenu>
     </Flex>
-  );
-}
-
-function SelectAgent(props) {
-  return (
-    <Box width="700px">
-      <Text mb={4} typography="h4" bold>
-        Connect a Resource
-      </Text>
-      <Text mb={2} bold>
-        Select Resource Type
-      </Text>
-      <Flex
-        alignItems="center"
-        css={`
-          margin: 0 auto;
-        `}
-      >
-        <AgentButton
-          as="button"
-          onClick={() => props.onClick('node')}
-          disabled={props.agentKind === 'node'}
-        >
-          <Icons.Apple fontSize="50px" my={2} />
-          <Text typography="h6" bold mt={1}>
-            Server
-          </Text>
-        </AgentButton>
-        <AgentButton
-          as="button"
-          onClick={() => props.onClick('app')}
-          disabled={props.agentKind === 'app'}
-        >
-          <Icons.Apple fontSize="50px" my={2} />
-          <Text typography="h6" bold mt={1}>
-            Application
-          </Text>
-        </AgentButton>
-      </Flex>
-    </Box>
   );
 }
 
@@ -255,95 +167,6 @@ function SideNavAgentConnect({ currentStep }) {
         </Box>
       </Box>
     </SideNavContainer>
-  );
-}
-
-function RequirementGathering(props: AgentStepProps) {
-  return (
-    <Box mt={5}>
-      <Text mb={2} bold>
-        General Requirement Gathering
-      </Text>
-      <Box width={500}>
-        Lorem ipsum dolor sit amet, consectetur adipiscing elit. Curabitur
-        lacinia tellus non velit porttitor, congue pretium quam luctus. Quisque
-        varius mi non purus consectetur, et dignissim odio cursus. Cras
-        tristique quis urna eget vehicula. Sed vehicula aliquam magna, in rutrum
-        nisl. Maecenas velit nisi, aliquam nec felis sed, ultricies euismod
-        tellus. Ut vel sem eget nisi tristique ullamcorper vel eget augue. Sed
-        imperdiet volutpat nisi vel mollis. Duis pulvinar, mauris sit amet
-        euismod dapibus, nibh felis dapibus ipsum, sit amet facilisis velit
-        nulla non risus. Aenean varius quam nulla. Sed a sem sagittis, gravida
-        ipsum quis, fringilla turpis.
-      </Box>
-      <ButtonPrimary width="182px" onClick={props.nextStep} mt={3}>
-        Proceed
-      </ButtonPrimary>
-    </Box>
-  );
-}
-
-function RequirementGatheringNode(props: AgentStepProps) {
-  return (
-    <Box mt={5}>
-      <Text mb={2} bold>
-        General Requirement Gathering Specific to Nodes
-      </Text>
-      <Box width={500}>Node stuff</Box>
-      <ButtonPrimary width="182px" onClick={props.nextStep} mt={3}>
-        Proceed
-      </ButtonPrimary>
-    </Box>
-  );
-}
-
-function TeleportInstallation(props: AgentStepProps) {
-  return (
-    <Box mt={5}>
-      <Text mb={2} bold>
-        General Teleport Installation
-      </Text>
-      <Box width={500}>
-        Lorem ipsum dolor sit amet, consectetur adipiscing elit. Curabitur
-        lacinia tellus non velit porttitor, congue pretium quam luctus.
-      </Box>
-      <ButtonSecondary mr={3} width="165px" onClick={props.prevStep} mt={3}>
-        Back
-      </ButtonSecondary>
-      <ButtonPrimary width="182px" onClick={props.nextStep} mt={3}>
-        Proceed
-      </ButtonPrimary>
-    </Box>
-  );
-}
-
-function AgentInitiation(props: AgentStepProps) {
-  return (
-    <Box mt={5}>
-      <Text mb={2} bold>
-        General Agent Installation
-      </Text>
-      <Box width={500}>
-        Lorem ipsum dolor sit amet, consectetur adipiscing elit. Curabitur
-        lacinia tellus non velit porttitor, congue pretium quam luctus. Quisque
-        varius mi non purus consectetur, et dignissim odio cursus. Cras
-        tristique quis urna eget vehicula. Sed vehicula aliquam magna, in rutrum
-        nisl. Maecenas velit nisi, aliquam nec felis sed, ultricies euismod
-        tellus. Ut vel sem eget nisi tristique ullamcorper vel eget augue. Sed
-        imperdiet volutpat nisi vel mollis. Duis pulvinar, mauris sit amet
-        euismod dapibus, nibh felis dapibus ipsum, sit amet facilisis velit
-        nulla non risus. Aenean varius quam nulla. Sed a sem sagittis, gravida
-        ipsum quis, fringilla turpis.
-      </Box>
-      <Box width={500}>
-        Lorem ipsum dolor sit amet, consectetur adipiscing elit. Curabitur
-        lacinia tellus non velit porttitor, congue pretium quam luctus. Quisque
-        varius mi non purus consectetur, et dignissim odio cursus.
-      </Box>
-      <ButtonSecondary mr={3} width="165px" onClick={props.prevStep}>
-        Back
-      </ButtonSecondary>
-    </Box>
   );
 }
 
@@ -405,75 +228,3 @@ export const MainContainer = styled.div`
   position: absolute;
   min-width: 1000px;
 `;
-
-const AgentButton = styled(Flex)(
-  props => `
-  align-items: center;
-  flex-direction: column;
-  transition: all 0.3s;
-  border-radius: 4px;
-  width: 160px;
-  border: none;
-  padding: 24px 32px;
-  margin-right: 16px;
-  margin-bottom: 16px;
-  background-color: ${props.theme.colors.primary.light};
-  &:disabled {
-    border: 2px solid ${props.theme.colors.secondary.main};
-    background: ${props.theme.colors.primary.lighter};
-    box-shadow: 0 4px 14px rgba(0, 0, 0, 0.56);
-  }
-  &:hover {
-    background: ${props.theme.colors.primary.lighter};
-  }
-  color: inherit;
-  cursor: pointer;
-  font-family: inherit;
-  text-align: center;
-`
-);
-
-type AgentStepProps = {
-  // nextStep increments the `currentStep` to go to the next step.
-  nextStep(): void;
-  // prevStep decrements the `currentStep` to go to the prev step.
-  prevStep(): void;
-  // updateResourceMeta updates the data as needed as we move through the step.
-  updateResourceMeta(meta: AgentMeta): void;
-};
-
-type AgentStepComponent = (props: AgentStepProps) => JSX.Element;
-
-// TODO (anyone): might need this to mimic the type token endpoint expects
-type AgentKind = 'app' | 'node';
-
-// AgentStep defines the order of steps in `connecting a agent (resource)`
-// that all agent kinds should share.
-//
-// The numerical enum value is used to determine which step the user is currently in,
-// which is also used as the `index value` to access array's values
-// for `agentStepTitles` and `agentViews`.
-enum AgentStep {
-  RequirementGather = 0,
-  TeleportInstall,
-  AgentInit,
-  RoleConfig,
-  ConnectResource,
-  TestResource,
-}
-
-// NodeMeta describes the fields that may be provided or required by user
-// when connecting a node.
-type NodeMeta = {
-  awsAccountId?: string;
-  awsArn?: string;
-};
-
-// AppMeta describes the fields that may be provided or required by user
-// when connecting a app.
-type AppMeta = {
-  name: string;
-  publicAddr: string;
-};
-
-type AgentMeta = NodeMeta | AppMeta;
