+++ packages/teleport/src/Discover/index.tsx
@@ -15,7 +15,7 @@
  */
 
 import React from 'react';
-import { Discover } from './Discover';
+import Discover from './Discover';
 import { DiscoverContext } from './discoverContext';
 import DiscoverContextProvider from './discoverContextProvider';
